Refactored voyage pool out of main and made separate object
