+++ src/main/java/setpartitioning/Data.java
@@ -47,12 +47,11 @@ public class Data {
         nbrVessels = Problem.getNumberOfVessels();
         vesselToNbrVoyages = new HashMap<>();
         for (int vIdx = 0; vIdx < nbrVessels; vIdx++) {
-            vesselToNbrVoyages.put(vIdx, Main.vesselToSequenceToCost.get(vIdx).size());
+            vesselToNbrVoyages.put(vIdx, VoyagePool.vesselToSequenceToCost.get(vIdx).size());
         }
     }
 
     public static void makeArrays() {
-        if (Main.vesselToSequenceToCost == null) throw new NullPointerException("VesselToSequenceToCost is null");
         makeOrdersArray();
         makeInstallationsArray();
         makeVesselToVoyageIndices();
@@ -81,7 +80,7 @@ public class Data {
             vesselToVoyageIdxToVoyage.put(vIdx, new HashMap<>());
             vesselToVoyageIdxToCost.put(vIdx, new HashMap<>());
             int voyageIdx = 0;
-            for (Map.Entry<List<Order>, Double> entry : Main.vesselToSequenceToCost.get(vIdx).entrySet()) {
+            for (Map.Entry<List<Order>, Double> entry : VoyagePool.vesselToSequenceToCost.get(vIdx).entrySet()) {
                 List<Order> voyage = entry.getKey();
                 double cost = entry.getValue();
                 vesselToVoyageIdxToVoyage.get(vIdx).put(voyageIdx, voyage);
